apps/etl/requirements.txt
pandas>=2.2
SQLAlchemy>=2.0
psycopg[binary]>=3.2
python-dotenv>=1.0